chore: move classes from new ruddersdkreactnative directory to the existing react.android
